microservices/api-gateway/requirements.txt
Flask==3.1.1
flask-cors==6.0.1
requests==2.31.0
gunicorn==23.0.0
PyJWT==2.10.1